+++ db/360/SceSysmem.yml
@@ -453,7 +453,6 @@ modules:
           ksceKernelGetPaddrPairForLargePage: 0x32257A24
           ksceKernelGetPaddrPairForSmallPage: 0xB3575090
           ksceKernelGetPhysicalMemoryType: 0x0AAA4FDD
-          ksceKernelGetPidContext: 0x2ECF7944
           ksceKernelGetUidClass: 0x85336A1C
           ksceKernelIsPaddrWithinSameSectionForUid: 0xF4AD89D8
           ksceKernelKernelUidForUserUidForClass: 0x184172B1
@@ -492,7 +491,6 @@ modules:
           ksceKernelStrncpyFromUser: 0xDB3EC244
           ksceKernelStrncpyToUser: 0x80BD6FEB
           ksceKernelStrnlenUser: 0xB429D419
-          ksceKernelSwitchPidContext: 0x2D711589
           ksceKernelSwitchVmaForPid: 0x6F2ACDAE
           ksceKernelUnmapMemBlock: 0xFFCD9B60
           ksceKernelUserMap: 0x278BC201
