+++ include/vitasdkkern.h
@@ -56,9 +56,9 @@
 #include <psp2kern/kernel/uart.h>
 #include <psp2kern/kernel/utils.h>
 #include <psp2kern/kernel/kbl/kbl.h>
+#include <psp2kern/kernel/sysmem/address_space.h>
 #include <psp2kern/kernel/sysmem/data_transfers.h>
 #include <psp2kern/kernel/sysmem/heap.h>
-#include <psp2kern/kernel/sysmem/mmu.h>
 #include <psp2kern/kernel/sysmem/uid_class.h>
 #include <psp2kern/kernel/sysmem/uid_guid.h>
 #include <psp2kern/kernel/sysmem/uid_puid.h>
